Move payload models to `models.remote` package
