+++ shared/src/androidMain/kotlin/dev/sasikanth/rss/reader/network/AndroidFeedParser.kt
@@ -16,7 +16,7 @@
 package dev.sasikanth.rss.reader.network
 
 import android.util.Xml
-import dev.sasikanth.rss.reader.models.FeedPayload
+import dev.sasikanth.rss.reader.models.remote.FeedPayload
 import dev.sasikanth.rss.reader.network.FeedParser.Companion.ATOM_TAG
 import dev.sasikanth.rss.reader.network.FeedParser.Companion.RSS_TAG
 import dev.sasikanth.rss.reader.utils.XmlParsingError
