+++ shared/src/androidMain/kotlin/dev/sasikanth/rss/reader/network/Parser.kt
@@ -15,7 +15,7 @@
  */
 package dev.sasikanth.rss.reader.network
 
-import dev.sasikanth.rss.reader.models.FeedPayload
+import dev.sasikanth.rss.reader.models.remote.FeedPayload
 import org.xmlpull.v1.XmlPullParser
 
 internal abstract class Parser {
