+++ shared/src/commonMain/kotlin/dev/sasikanth/rss/reader/models/remote/FeedPayload.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.sasikanth.rss.reader.models
+package dev.sasikanth.rss.reader.models.remote
 
 data class FeedPayload(
   val name: String,
