+++ shared/src/commonMain/kotlin/dev/sasikanth/rss/reader/network/FeedFetcher.kt
@@ -15,7 +15,7 @@
  */
 package dev.sasikanth.rss.reader.network
 
-import dev.sasikanth.rss.reader.models.FeedPayload
+import dev.sasikanth.rss.reader.models.remote.FeedPayload
 import io.ktor.client.HttpClient
 import io.ktor.client.request.get
 import io.ktor.client.statement.bodyAsText
