+++ shared/src/commonMain/kotlin/dev/sasikanth/rss/reader/network/FeedParser.kt
@@ -15,7 +15,7 @@
  */
 package dev.sasikanth.rss.reader.network
 
-import dev.sasikanth.rss.reader.models.FeedPayload
+import dev.sasikanth.rss.reader.models.remote.FeedPayload
 
 internal interface FeedParser {
