+++ shared/src/iosMain/kotlin/dev/sasikanth/rss/reader/network/IOSFeedParser.kt
@@ -17,8 +17,8 @@
 
 package dev.sasikanth.rss.reader.network
 
-import dev.sasikanth.rss.reader.models.FeedPayload
-import dev.sasikanth.rss.reader.models.PostPayload
+import dev.sasikanth.rss.reader.models.remote.FeedPayload
+import dev.sasikanth.rss.reader.models.remote.PostPayload
 import dev.sasikanth.rss.reader.network.FeedParser.Companion.ATOM_TAG
 import dev.sasikanth.rss.reader.network.FeedParser.Companion.RSS_TAG
 import dev.sasikanth.rss.reader.network.FeedParser.Companion.imageTags
